Split up xtask commands
